+++ MusicStreamingWebApp/MusicStreaming.html
@@ -52,6 +52,7 @@
             </fieldset>
         </div>
     </div>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/3.0.3/socket.io.js" integrity="sha512-Jr0UIR/Q8MUX+93zjDOhuDUKLqJZObtwpkLJQcR9qMaLgL0thet39IORuavUaZFkZ8a4ktrUsKPM9mf5LWMduA==" crossorigin="anonymous"></script>
     <script src="js\MusicStreaming.js"></script>
 </body>
